refactor: organize files into components and utils
