+++ src/utils/timerUtils.ts
