+++ tests/sound.test.mjs
@@ -3,7 +3,7 @@ import test from 'node:test';
 
 test('playSound calls react-sounds with success_chime', async () => {
   const calls = [];
-  const mod = await import('../src/sound.ts');
+  const mod = await import('../src/utils/sound.ts');
   mod.setPlaySoundImpl(async (...args) => {
     calls.push(args);
   });
